create_py_project/stubs/template-drf-and-auth/api/urls.py
from django.urls import include, path
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
    TokenVerifyView
)

urlpatterns = [
    path('auth/user/', include('accounts.urls')),
    path('token/', TokenObtainPairView.as_view(), name='token'),
    path('token/verify/', TokenVerifyView.as_view(), name='verify_token'),
    path('token/refresh/', TokenRefreshView.as_view(), name='refresh_token'),
]
