create_py_project/stubs/template-drf-and-auth/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.1
django-cors-headers==3.13.0
django-filter==22.1
django-jazzmin==2.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
importlib-metadata==5.0.0
Markdown==3.4.1
PyJWT==2.5.0
pytz==2022.4
sqlparse==0.4.3
zipp==3.8.1
